% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeSimulatedData.R
\name{AnalyzeSimulatedData}
\alias{AnalyzeSimulatedData}
\title{Analyze simulation reslts}
\usage{
AnalyzeSimulatedData(
  Sim,
  Forward,
  Reverse,
  pi0 = NULL,
  psm.fdr.bin = c(0, 0.01, 0.05, seq(0.1, 1, by = 0.1))
)
}
\arguments{
\item{Sim}{Simulation output from \code{SimulateDataPaper}}

\item{Forward}{Target search results from \code{PerformSearch}}

\item{Reverse}{Decoy search results from \code{PerformSearch}}

\item{pi0}{If known, the fraction of simulated spectra whose generating analyte does not lie in the target database. Detault is \code{NULL}, i.e. unknown}

\item{psm.fdr.bin}{PSM-fdr bins to determine PSM-fdr calibration}
}
\value{
A list \item{S.F}{Bayes factor score for target database} \item{S.R}{Bayes factor score for reverse database} \item{SEE.F}{Score-ordering error rate} \item{Sfdr}{DI-fdr} \item{PSMfdr}{PSM-fdr} \item{pi0.hat}{Estimate for pi0} \item{pvalues}{p-values for the null hypothesis that the spectrum was not generated by an analyte in the target database} \item{Pep.F}{Inferred target peptide} \item{Correct}{Is \code{Pep.F} correct} \item{PSMfdr.bins}{PSM-fdr calibration results}
}
\description{
Analyze results derived from simulated data.
}
