% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeIsotope_msms.R
\name{Deisotope}
\alias{Deisotope}
\title{Deisotope MS/MS spectra}
\usage{
Deisotope(
  mzml.file.path,
  n_cores = NULL,
  ppm = 20,
  ppm.filter = 40,
  ppm.precursor = 20,
  remove.precursor = F,
  max.addition = 2,
  min.ions = 20,
  min.chi = 0.26,
  write.mzml = F
)
}
\arguments{
\item{mzml.file.path}{A character string containing the mzML or mzXML file path}

\item{n_cores}{Number of cores. Defaults to the total number of available cores minus 1.}

\item{ppm}{Fragment isotope mass error, in the ppm scale. Detaults to 20.}

\item{ppm.filter}{The software will collapse any peaks <=\code{ppm.filter}ppm apart into 1 peak. Defaults to 40.}

\item{remove.precursor}{Should the precursor be removed. Defaults to \code{F}.}

\item{write.mzml}{Should an mzML file be returned. Defaults to \code{F}.}
}
\value{
If \code{write.mzml} is set to \code{T}, returns an mzML file in the same directory as \code{mzml.file.path}. Otherwise, it returns a list with the following entries: \item{Headers}{A data frame of headers for each MS and MS/MS spectrum.} \item{Spectra}{A list containing the MS and MS/MS spectra. The outputed MS/MS spectra contain 3 columns: m/z, intensity, and inferred charge.}
}
\description{
Deisotope MS/MS spectra using a modified chi squared statistic scoring function.
}
