% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScorePSMs.R
\name{PerformSearch}
\alias{PerformSearch}
\title{Search MS/MS spectra with MSeQUiP}
\usage{
PerformSearch(
  Spectra,
  Headers = NULL,
  data.F,
  clean.database = F,
  match.charge = T,
  append.file = "_Train.Q9.lowpH",
  scale = c("Quant90", "base-peak", "none"),
  path.spline = "<path to>/NoiseObserved",
  rt.path = "<path to>/RetentionTime",
  byInt.path = "<path to>/BYInt",
  byMA.path = "<path to>/BYMassAccuracy",
  byObs.path = "<path to>/BYObserved",
  NoiseInt.path = "<path to>/NoiseInt",
  Noise.MA.path = "<path to>/NoiseMassAccuracy",
  Int0.path = NULL,
  ppm.precursor = 10,
  min.peaks = 20,
  delta.ppm = 20,
  delta.ppm.prec = 20,
  scale.obs = NA,
  include.scale = T,
  append.noise = NULL,
  include.pep.IDs = T
)
}
\arguments{
\item{Spectra}{A list of MS and MS/MS spectra. Can be obtained using mzR, or \code{Deisotope}.}

\item{Headers}{A data frame of headers. Can be obtained using mzR, or \code{Deisotope}.}

\item{data.F}{Spectral library. This can be obtained from Prosit.}

\item{clean.database}{Should the spectral library be cleaned. Defaults to \code{F}.}

\item{match.charge}{Should inferred fragment charges obtained from \code{Deisotope} be used. Defaults to \code{T}.}

\item{append.file}{Name of training parameters to use. Detaults to \code{"_Train.Q9.lowpH"}.}

\item{scale}{Type of intensity normalization to use. Can be one of \code{"Quant90"} (normalize by the 0.9 quantile), \code{"base-peak"} (normalize by the base peak), or \code{"none"} (no normalization). Defaults to \code{"Quant90"}.}

\item{path.spline}{Path to parameters to define lambda(m), the probability a noise peak was generated around a predicted m/z m}

\item{rt.path}{Path to retention time parameters.}

\item{byInt.path}{Path to signal b/y intensity parameters.}

\item{byMA.path}{Path to signal mass accuracy parameters.}

\item{byObs.path}{Path to parameters determine the probability a b/y ion is observed.}

\item{NoiseInt.path}{Path to parameters describing the noise intensity.}

\item{Noise.MA.path}{Path to parameters describing the noise mass accuracy.}

\item{ppm.precursor}{Precursor mass accuracy, on the ppm scale. Defaults to 10.}

\item{min.peaks}{Minimum number of peaks in the MS/MS spectrum. Defaults to 20.}

\item{delta.ppm}{MS/MS fragment mass accuracy, on the ppm scale. Defaults to 20.}

\item{delta.ppm.prec}{Maximum difference between the precursors observed in MS and MS/MS spectra.}

\item{scale.obs}{The intensity scale, if available. The default of \code{NA} will almost always be appropriate.}

\item{include.scale}{If \code{scale.obs} is available, should it be used. Defaults to \code{T}.}

\item{append.noise}{Name of training parameters to use. Detaults to \code{append.file}.}

\item{include.pep.IDs}{Should peptide IDs be included in the results. Defaults to \code{T}.}
}
\value{
A list. \item{MS2.indices}{The indices of \code{Spectra} that are MS/MS spectra.} \item{Results}{A list. List element \code{FM} contains the results, and is a list of matrices giving the MSeQUiP scores and other statistics for each scored peptide. The sum of the first 4 columns gives the log Bayes factor for the peptide. The element \code{RM} is NA.}
}
\description{
Computes MSeQUiP Bayes factors for each candidate peptide
}
