% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateNull.R
\name{SimulateDataPaper}
\alias{SimulateDataPaper}
\title{Simulate data from the paper}
\usage{
SimulateDataPaper(ProcessedDataPath, n.null, t.df.by = 4, list.train.paths)
}
\arguments{
\item{ProcessedDataPath}{Path to processed data.}

\item{n.null}{Number of spectra to simulate subject to database incompleteness errors.}

\item{t.df.by}{Log-signal intensity errors are drawn from a t distribution with \code{t.df.by} degrees of freedom.}
}
\value{
A list. \item{Spectra.sim}{A list of simulated spectra. It is assumed each MS/MS peak's charge is unknown in these simulations.} \item{Headers.sim}{Headers for the simulated spectra} \item{Pep}{Generating peptide} \item{Charge}{Precursor charge} \item{Scale}{The intensity scale for each simulated spectrum}
}
\description{
Simulate MS/MS spectra subject to score-ordering and database incompleteness errors. The former contain simulated signal and real noise peaks, and the latter are real spectra whose precursors have been shifted +/-10m/z.
}
